Cache all 4 corners when not doing offscreen render
When using cache only for 2 left corners like in the previous commit, there can be artifacts on right corners in cases with dark background on the left. To avoid that, let's cache shadows for all 4 corners when no offscreen render is performed. For cases, where an offscreen render is needed, only 2 left corners will be cached to gain some performance. Their copies will be stored for the right ones and fliped horizontally in the shader when rendered. Since the render is done to a white texture, there will be no artifacts during restoring the shadows.